sdk: store contracts info in state
Before, when using udcAddress as contract's entrypoint, the discovery
had to be performed every session, as it wasn't stored, and also because
this contract's discovery was needed to get TNR addr and find state.
Now, we use UDC address to index state, and persist the whole
discovered contractsInfo in it, so we reuse from state after first init.